git: check for transient errors on pull (#59)
* git: check for transient errors on pull

if we get 'connection refused', don't assume our git repo is borked.
This is done by checking the error message of git - which is not great.

Better would be to 'git fetch' and 'git merge', but unlike 'git pull'
does cmds don't report what files are changed? (Or do they). Otherwise
we could ignore 'git fetch' errors and only delete the repo on 'git
merge' failures.

Signed-off-by: Miek Gieben <miek@miek.nl>

* Retry initial check out as well

Use same duration and loop as for normal tracking routine.

Signed-off-by: Miek Gieben <miek@miek.nl>

---------

Signed-off-by: Miek Gieben <miek@miek.nl>